refactor(RAbbitMQ): fix build

+++ AMQP-CPP/example/2/SmartBuffer.h
@@ -39,24 +39,24 @@ class SmartBuffer
         return _buffer.data();
     }
 
-    char* shrink( unsigned int elementsToRemove )
+    char* shrink( uint64_t elementsToRemove )
     {
-        elementsToRemove = std::min( elementsToRemove,(unsigned int)_buffer.size() );
-        _buffer.erase ( _buffer.begin(), _buffer.begin()+elementsToRemove );
+        elementsToRemove = ( uint64_t )std::min( elementsToRemove, _buffer.size() );
+        _buffer.erase ( _buffer.begin(), _buffer.begin() + elementsToRemove );
         read( _eventFD, & elementsToRemove, sizeof( elementsToRemove ) );
         return _getBuffer();
     }
 
-    void append( const char* data, unsigned int size )
+    void append( const char* data, uint64_t size )
     {
-        _buffer.insert( _buffer.end(), data, data+size );
+        _buffer.insert( _buffer.end(), data, data + size );
         write( _eventFD, & size, sizeof( size ) );
     }
 
     void clear()
     {
         _buffer.clear();
-        ssize_t i = _buffer.size();
+        uint64_t i = _buffer.size();
         read( _eventFD, & i, sizeof( i ) );
     }
